fix windows build (not working yet; select and pipe dont play well)

+++ test/sockpuppet_tcp_async_test.cpp
@@ -11,7 +11,7 @@ struct Server
 {
   SocketTcpAsyncServer server;
   SocketDriver &driver;
-  unsigned int bytesReceived;
+  size_t bytesReceived;
   std::deque<SocketTcpAsyncClient> serverHandlers;
   std::mutex mtx;
